Revise README slightly
One small correction: the original ed25519.py is not what DJB calls
the reference implementation.  That would be the C implementation in
SUPERCOP, which is referred to in the paper (and which PyNaCl
provides.)

Also expand a bit on the contrast between this implementation and the
one provided by PyNaCl.